add release action in buttons manager

+++ c/hw_drivers/gpio/gpio_i94xxx/src/gpio_i94xxx.h
@@ -25,7 +25,7 @@ struct gpio_i94xxx_config_t
 
 struct gpio_i94xxx_runtime_t
 {
-	uint8_t  pin_mask;
+	uint16_t  pin_mask;
 	uint8_t  pin_bitwise_curr_values[2];
 	uint8_t  pin_bitwise_idle_values[2];
 };
